+++ src/ray/gcs/gcs_server/default_job_info_access_handler.cc
@@ -3,13 +3,13 @@
 namespace ray {
 namespace rpc {
 void DefaultJobInfoAccessHandler::HandleAddJob(
-    const rpc::GcsJobInfo &request, rpc::AddJobReply *reply,
+    const rpc::AddJobRequest &request, rpc::AddJobReply *reply,
     rpc::SendReplyCallback send_reply_callback) {
   RAY_LOG(DEBUG) << "Received new job ...";
 }
 
 void DefaultJobInfoAccessHandler::HandleMarkJobFinished(
-    const rpc::FinishedJob &request, rpc::MarkJobFinishedReply *reply,
+    const rpc::MarkJobFinishedRequest &request, rpc::MarkJobFinishedReply *reply,
     rpc::SendReplyCallback send_reply_callback) {
   RAY_LOG(DEBUG) << "Mark job as finished ...";
 }
