+++ src/ray/gcs/gcs_server/default_job_info_access_handler.h
@@ -1,15 +1,16 @@
 #pragma once
-#include "ray/rpc/gcs_server/job_info_access_server.h"
+#include "ray/rpc/gcs_server/gcs_rpc_server.h"
 
 namespace ray {
 namespace rpc {
 
 class DefaultJobInfoAccessHandler : public rpc::JobInfoAccessHandler {
  public:
-  void HandleAddJob(const GcsJobInfo &request, AddJobReply *reply,
+  void HandleAddJob(const AddJobRequest &request, AddJobReply *reply,
                     SendReplyCallback send_reply_callback) override;
 
-  void HandleMarkJobFinished(const FinishedJob &request, MarkJobFinishedReply *reply,
+  void HandleMarkJobFinished(const MarkJobFinishedRequest &request,
+                             MarkJobFinishedReply *reply,
                              SendReplyCallback send_reply_callback) override;
 };
